fix(confignet): resolve CI lint failures and Windows test skip
Move Unix-socket-specific tests to confignet_unix_test.go with
//go:build !windows constraint to fix Windows CI failures.
Fix lint issues: gosec G306 (0o644→0o600), testifylint require-error,
usetesting nolint directives for short socket paths, and gofumpt
trailing whitespace.
