+++ config/confignet/confignet_test.go
@@ -7,10 +7,7 @@ import (
 	"context"
 	"errors"
 	"net"
-	"os"
-	"path/filepath"
 	"strings"
-	"syscall"
 	"testing"
 	"time"
 
@@ -202,167 +199,11 @@ func Test_TransportType_UnmarshalText(t *testing.T) {
 	require.Error(t, err)
 }
 
-func Test_removeStaleSocket(t *testing.T) {
-	t.Parallel()
-	t.Run("path does not exist", func(t *testing.T) {
-		t.Parallel()
-		err := removeStaleSocket(filepath.Join(t.TempDir(), "nonexistent.sock"))
-		assert.NoError(t, err)
-	})
-
-	t.Run("path is a regular file", func(t *testing.T) {
-		t.Parallel()
-		dir := t.TempDir()
-		path := filepath.Join(dir, "regular.txt")
-		require.NoError(t, os.WriteFile(path, []byte("data"), 0o644))
-
-		err := removeStaleSocket(path)
-		assert.ErrorContains(t, err, "not a socket")
-		// File should still exist.
-		_, statErr := os.Stat(path)
-		assert.NoError(t, statErr)
-	})
-
-	t.Run("path is a stale socket", func(t *testing.T) {
-		t.Parallel()
-		// Use os.MkdirTemp with an empty base so the OS chooses a short path,
-		// avoiding the 104-character Unix socket path limit on macOS/BSD.
-		dir, err := os.MkdirTemp("", "confignet-sock-test-*")
-		require.NoError(t, err)
-		t.Cleanup(func() { os.RemoveAll(dir) })
-		path := filepath.Join(dir, "stale.sock")
-		// Create a socket file directly via syscall so it is not auto-removed
-		// when closed (Go's net.Listener.Close removes socket files on some OSes).
-		fd, err := syscall.Socket(syscall.AF_UNIX, syscall.SOCK_STREAM, 0)
-		require.NoError(t, err)
-		err = syscall.Bind(fd, &syscall.SockaddrUnix{Name: path})
-		require.NoError(t, err)
-		require.NoError(t, syscall.Close(fd))
-		// Socket file should still exist after closing the fd.
-		_, err = os.Stat(path)
-		require.NoError(t, err)
-
-		err = removeStaleSocket(path)
-		assert.NoError(t, err)
-		// Socket file should be removed.
-		_, err = os.Stat(path)
-		assert.True(t, os.IsNotExist(err))
-	})
-
-}
-
 func Test_isAbstractSocket(t *testing.T) {
 	t.Parallel()
 	assert.True(t, isAbstractSocket("@abstract"))
 	assert.False(t, isAbstractSocket("/var/run/test.sock"))
 	assert.False(t, isAbstractSocket(""))
-}
-
-func Test_unixListener_Close(t *testing.T) {
-	t.Parallel()
-	dir, err := os.MkdirTemp("", "confignet-test")
-	require.NoError(t, err)
-	t.Cleanup(func() { os.RemoveAll(dir) })
-	path := filepath.Join(dir, "listener.sock")
-
-	ln, err := net.Listen("unix", path)
-	require.NoError(t, err)
-
-	// Socket file exists after listen.
-	_, err = os.Stat(path)
-	require.NoError(t, err)
-
-	uln := &unixListener{Listener: ln, path: path}
-	require.NoError(t, uln.Close())
-
-	// Socket file should be removed after close.
-	_, err = os.Stat(path)
-	assert.True(t, os.IsNotExist(err))
-}
-
-func TestAddrConfig_Listen_UnixRemovesStaleSocket(t *testing.T) {
-	t.Parallel()
-	dir, err := os.MkdirTemp("", "confignet-test")
-	require.NoError(t, err)
-	t.Cleanup(func() { os.RemoveAll(dir) })
-	path := filepath.Join(dir, "stale.sock")
-
-	// Create a stale socket using syscall (macOS removes socket on net.Listener.Close).
-	fd, sysErr := syscall.Socket(syscall.AF_UNIX, syscall.SOCK_STREAM, 0)
-	require.NoError(t, sysErr)
-	require.NoError(t, syscall.Bind(fd, &syscall.SockaddrUnix{Name: path}))
-	require.NoError(t, syscall.Close(fd))
-	_, sysErr = os.Stat(path)
-	require.NoError(t, sysErr)
-
-	// Listen should succeed despite stale socket.
-	na := &AddrConfig{
-		Endpoint:  path,
-		Transport: TransportTypeUnix,
-	}
-	ln, err := na.Listen(context.Background())
-	require.NoError(t, err)
-	t.Cleanup(func() { require.NoError(t, ln.Close()) })
-}
-
-func TestAddrConfig_Listen_UnixRefusesToRemoveNonSocket(t *testing.T) {
-	t.Parallel()
-	dir, err := os.MkdirTemp("", "confignet-test")
-	require.NoError(t, err)
-	t.Cleanup(func() { os.RemoveAll(dir) })
-	path := filepath.Join(dir, "regular.txt")
-	require.NoError(t, os.WriteFile(path, []byte("data"), 0o644))
-
-	na := &AddrConfig{
-		Endpoint:  path,
-		Transport: TransportTypeUnix,
-	}
-	_, err = na.Listen(context.Background())
-	assert.ErrorContains(t, err, "not a socket")
-}
-
-func TestAddrConfig_Listen_UnixSocketPermissions(t *testing.T) {
-	t.Parallel()
-	dir, err := os.MkdirTemp("", "confignet-test")
-	require.NoError(t, err)
-	t.Cleanup(func() { os.RemoveAll(dir) })
-	path := filepath.Join(dir, "perms.sock")
-
-	na := &AddrConfig{
-		Endpoint:  path,
-		Transport: TransportTypeUnix,
-	}
-	ln, err := na.Listen(context.Background())
-	require.NoError(t, err)
-	t.Cleanup(func() { require.NoError(t, ln.Close()) })
-
-	fi, err := os.Stat(path)
-	require.NoError(t, err)
-	assert.Equal(t, socketFileMode|os.ModeSocket, fi.Mode())
-}
-
-func TestAddrConfig_Listen_UnixSocketCloseRemovesFile(t *testing.T) {
-	t.Parallel()
-	dir, err := os.MkdirTemp("", "confignet-test")
-	require.NoError(t, err)
-	t.Cleanup(func() { os.RemoveAll(dir) })
-	path := filepath.Join(dir, "cleanup.sock")
-
-	na := &AddrConfig{
-		Endpoint:  path,
-		Transport: TransportTypeUnix,
-	}
-	ln, err := na.Listen(context.Background())
-	require.NoError(t, err)
-
-	_, err = os.Stat(path)
-	require.NoError(t, err)
-
-	require.NoError(t, ln.Close())
-
-	// Socket file should be removed after close.
-	_, err = os.Stat(path)
-	assert.True(t, os.IsNotExist(err))
 }
 
 func Test_AddrConfig_isUnixTransport(t *testing.T) {
